Updated tests in order to compile with new protocols
