+++ SplitTests/Service/MySegments/MySegmentsBgSyncWorkerTest.swift
@@ -39,7 +39,7 @@ class MySegmentsBgSyncWorkerTest: XCTestCase {
 
     func testNoSuccess() {
 
-        mySegmentsFetcher.httpError = HttpError.clientRelated(code: -1)
+        mySegmentsFetcher.httpError = HttpError.clientRelated(code: -1, internalCode: -1)
         mySegmentsFetcher.allSegments = [["s1", "s2"]]
         mySegmentsSyncWorker.execute()
